+++ src/packages/@library/icons/index.ts
@@ -39,3 +39,4 @@ export { default as SettingIcon } from "./Setting.svelte";
 export { default as HelpCenterIcon } from "./HelpCenter.svelte";
 export { default as InformationIcon } from "./Information.svelte";
 export { default as LockIcon } from "./Lock.svelte";
+export { default as MoreOptions } from "./MoreOptions.svelte";
