+++ src/packages/@teams/features/index.ts
@@ -4,3 +4,4 @@ export * from "./create-team";
 export * from "./team-invite";
 export * from "./team-members";
 export * from "./team-settings";
+export * from "./leave-team";
