+++ src/packages/@teams/features/team-explorer/layout/TeamExplorer.svelte
@@ -17,9 +17,8 @@
   import { WorkspaceListView } from "../components";
   import WorkspaceGridView from "../components/workspace-grid-view/WorkspaceGridView.svelte";
   import { TeamMembers, TeamSettings } from "@teams/features";
-  import { CrossIcon } from "@library/icons";
-  import MoreOptions from "@workspaces/features/tab-bar/assets/MoreOptions.svelte";
-  import Tooltip from "@library/ui/tooltip/Tooltip.svelte";
+  import { CrossIcon , MoreOptions } from "@library/icons";
+  import { Tooltip } from "@library/ui";
   import { Dropdown } from "@library/ui";
 
   /**
